api/pyproject.toml
[tool.poetry]
name = "hittite-transcription"
version = "0.1.0"
description = ""
authors = ["yustoris <yustoris@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
fastapi = "^0.68.0"
uvicorn = "^0.14.0"
lark = "^0.11.3"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
